refactor(project-management): split work service modules
Turn work_service/mod.rs into a stable facade and move creation, lifecycle, mutation, notes, relations, idempotency, run-review, bootstrap, error, and test-support responsibilities into focused modules. Re-export the existing public and crate-visible API so downstream callers remain unchanged.

Verification: cargo fmt --check, cargo check --all-targets, cargo clippy --all-targets with warnings denied, and all 619 project_management library tests.

Pre-commit hook ran. Total eslint: 0, total circular: 0
